@@ -19,26 +19,25 @@
 		"@auth/sveltekit": "^0.14.0",
 		"@ltd/j-toml": "^1.38.0",
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"autoprefixer": "^10.4.16",
+		"@sveltejs/kit": "^2.22.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
+		"autoprefixer": "^10.4.21",
 		"eslint": "^8.55.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
 		"jszip": "^3.10.1",
-		"monaco-editor": "^0.44.0",
+		"monaco-editor": "^0.52.2",
 		"mongodb": "^6.5.0",
 		"postcss": "^8.4.32",
 		"prettier": "^3.1.0",
 		"prettier-plugin-svelte": "^3.1.2",
 		"prettier-plugin-tailwindcss": "^0.5.7",
 		"sass": "^1.71.1",
-		"svelte": "^4.2.8",
+		"svelte": "^5.34.7",
 		"svelte-check": "^3.6.2",
-		"svelte-dialogs": "github:asm80/svelte-dialogs",
 		"tslib": "^2.6.2",
 		"typescript": "^5.3.2",
-		"vite": "^5.0.0",
+		"vite": "^5.4.19",
 		"vite-plugin-devtools-json": "^0.2.0",
 		"vitest": "^1.0.0"
 	},
