@@ -1,7 +1,7 @@
 <script>
 	//import '../app.css';
-	import '../styles.scss';
-	import '../assets/fontawesome/css/all.css';
+	import '../../styles.scss';
+	import '../../assets/fontawesome/css/all.css';
 </script>
 
 <slot />
